# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Expenses tracker API"
  },
  "securityDefinitions": {
    "BasicAuth": {
      "name": "Authorization",
      "in": "header",
      "type": "basic",
      "description": "basic http auth header"
    }
  },
  "security": [
    {
      "BasicAuth": []
    }
  ],
  "paths": {
    "/expenses/user/{user_id}": {
      "parameters": [
        {
          "name": "user_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Expenses"
        ],
        "summary": "get all user expenses",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/expense/user/{user_id}": {
      "parameters": [
        {
          "name": "user_id",
          "in": "path",
          "required": true,
          "type": "string"
        },
        {
          "name": "expense",
          "in": "body",
          "required": true,
          "schema": {
          }
        }
      ],
      "post": {
        "tags": [
          "Expense"
        ],
        "summary": "post a new Expense",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/expenses-statistics/user/{user_id}/year/{year}": {
      "parameters": [
        {
          "name": "user_id",
          "in": "path",
          "required": true,
          "type": "string"
        },
        {
          "name": "year",
          "in": "path",
          "required": true,
          "type": "number"
        }
      ],
      "get": {
        "tags": [
          "Expenses-statistics"
        ],
        "summary": "get al expenses by year, and statistics: sum and count",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/expenses-statistics/user/{user_id}/year/{year}/month/{month}": {
      "parameters": [
        {
          "name": "user_id",
          "in": "path",
          "required": true,
          "type": "string"
        },
        {
          "name": "year",
          "in": "path",
          "required": true,
          "type": "number"
        },
        {
          "name": "month",
          "in": "path",
          "required": true,
          "type": "number"
        }
      ],
      "get": {
        "tags": [
          "Expenses-statistics"
        ],
        "summary": "get al expenses by month, and statistics: sum and count",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/expenses-statistics/user/{user_id}/category/{category}": {
      "parameters": [
        {
          "name": "user_id",
          "in": "path",
          "required": true,
          "type": "string"
        },
        {
          "name": "category",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Expenses-statistics"
        ],
        "summary": "get al expenses by category, and statistics: sum and count",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user": {
      "parameters": [
        {
          "name": "user",
          "in": "body",
          "required": true,
          "schema": {
          }
        }
      ],
      "post": {
        "tags": [
          "User"
        ],
        "summary": "post a new User",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}